[Module-3]Implement Clear Architecture for Domain.
