+++ LibraryApp/LibraryApp.Api/Controllers/LibraryController.cs
@@ -1,5 +1,7 @@
-﻿using LibraryApp.console.Domain;
-using LibraryApp.console.Services;
+﻿//using LibraryApp.console.Domain;
+//using LibraryApp.console.Services;
+using LibraryApp.Domain;
+using LibraryApp.Services;
 using Microsoft.AspNetCore.Mvc;
 
 namespace LibraryApp.Api.Controllers
@@ -24,17 +26,17 @@ namespace LibraryApp.Api.Controllers
         }
 
         [HttpPost("add-book")]
-        public IActionResult AddBook([FromBody] AddBookRequest request)
+        public IActionResult AddBook([FromBody] Book book)
         {
             // 1. Validate the request data.
-            if (string.IsNullOrWhiteSpace(request.Title) || string.IsNullOrWhiteSpace(request.Author))
+            if (book == null || string.IsNullOrWhiteSpace(book.Title) || string.IsNullOrWhiteSpace(book.Author))
             {
                 // Return a 400 Bad Request if the required fields are missing.
                 return BadRequest("Title and Author are required.");
             }
 
             // 2. Call the service to add the book.
-            var newBook = _service.AddBook(request.Title, request.Author, request.Pages);
+            var newBook = _service.AddBook(book.Title, book.Author, book.Pages);
 
             // 3. Return a 201 Created status code with the newly created book.
             // The Url.Action method helps generate the URL for the created resource.
