+++ LibraryApp/LibraryApp.Api/Program.cs
@@ -1,4 +1,5 @@
-using LibraryApp.console.Services;
+//using LibraryApp.console.Services;
+using LibraryApp.Services;
 
 var builder = WebApplication.CreateBuilder(args);
