+++ LibraryApp/LibraryApp.Domain/Book.cs
@@ -4,9 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LibraryApp.console.Domain
+namespace LibraryApp.Domain
 {
-    
     public class Book : LibraryItem
     {
         public string Author { get; }
@@ -14,7 +13,7 @@ namespace LibraryApp.console.Domain
         public int Pages { get; }
 
         //constructor chaining example
-        
+
         public Book(int id, string title, string author) : this(id, title, author, pages: 0) { }
         public Book(int id, string title, string author, int pages) : base(id, title)
         {
@@ -23,8 +22,8 @@ namespace LibraryApp.console.Domain
         }
 
         public override string GetInfo() =>
-        
+
              $"Book [Id={Id}, Title={Title}, Author={Author}, Pages={Pages}, IsBorrowed={IsBorrowed}]";
-        
+
     }
 }
