+++ LibraryApp/LibraryApp.Domain/Magazine.cs
@@ -1,4 +1,10 @@
-﻿namespace LibraryApp.console.Domain
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LibraryApp.Domain
 {
     public sealed class Magazine : LibraryItem
     {
@@ -13,5 +19,4 @@
         public override string GetInfo()
             => $"[Magazine] {Title} - Issue #{IssueNumber} ({Publisher})";
     }
-
 }
