+++ LibraryApp/LibraryApp.Services/ILibraryService.cs
@@ -1,6 +1,6 @@
-﻿using LibraryApp.console.Domain;
+﻿using LibraryApp.Domain;
 
-namespace LibraryApp.console.Services
+namespace LibraryApp.Services
 {
     public interface ILibraryService
     {
